pkggrps: move sysconfig-settings-ui to graphical
The sysconfig-settings-ui package provides the uixml and other elements
needed for syscfg (MAX, HWCU) to affect a toggle button for the embeded
UI. This toggle is only meaningful for targets with graphical drivers
and in the context of the graphical packagegroup.

In some cases (SNAC), we want to disable the UI toggle, even though the
device has graphical capability. But opkg will complain that removing
the -ui package invalidates the pkggrp-ni-runmode dependencies.

Move the package to pkggrp-ni-graphical, because it makes more logical
sense and stops opkg from complaining.

Signed-off-by: Alex Stewart <alex.stewart@emerson.com>