pkggrp-ni-extra: simplify legacy branch logic
A section of the extra package feed previously only built on x64 distro
builds, when X11 was enabled. This was once meaningful when we had
multiple concurrent distro flavors (legacy and NXG) and concurrent
architectures (ARM32 and x64).

We now only build one flavor and x64 only. So collapse this branch logic
for simplicity.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>