sysconfig-settings: subpackage consoleout content
The sysconfig-settings package includes uixml and other elements that
enable MAX and HWCU to control the consoleout firmware setting, and
therefore control whether or not a console TTY is provided at boot.

The NILRT SNAC configuration prohibits console access as a matter of
policy. In order to remove a tempting toggle for SNAC users, the
consoleout content needs to be cleanly separated from the
sysconfig-settings remainder.

Split the consoleout contents into their own sysconfig-settings-console
subpackage.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>